riconoscitore_flask/requirements.txt
Flask==3.1.1
requests>=2.25.1
beautifulsoup4>=4.9.3
torch>=1.13.0
torchvision==0.22.1
Pillow>=8.0.0
opencv-contrib-python==4.12.0.88
matplotlib==3.10.3
nltk==3.9.1
pip-autoremove==0.10.0
pipdeptree==2.26.1
PyWavelets==1.8.0
scikit-image==0.25.2
scikit-learn==1.7.0
transformers==4.53.2
wheel==0.45.1
